+++ NEWS.md
@@ -1,3 +1,7 @@
+0.3.0
+* ALTREP-aware
+* `length()` and `[` implemented for `tdigest` objects
+
 0.2.0
 * Added input vaildity checks
 * Added `quantile()` function S3 implementation for `tdigest` objects